Change float type to f64

+++ src/lib.rs
@@ -9,14 +9,14 @@ pub const RELU: fn(f32) -> f32 = |x| x.max(0.0);
 
 pub struct NeuralNet {
     layers: Vec<Layer>,
-    weights: Vec<DMatrix<f32>>,
-    biases: Vec<DMatrix<f32>>,
-    activation: fn(f32) -> f32,
+    weights: Vec<DMatrix<f64>>,
+    biases: Vec<DMatrix<f64>>,
+    activation: fn(f64) -> f64,
 }
 
 impl NeuralNet {
     // TODO: Add check to see if more than 1 layer was supplied
-    pub fn new(node_counts: Vec<usize>, activation: fn(f32) -> f32, rng: &mut impl Rng) -> Self {
+    pub fn new(node_counts: Vec<usize>, activation: fn(f64) -> f64, rng: &mut impl Rng) -> Self {
         Self {
             layers: node_counts.iter().map(|c| Layer::new(*c)).collect(),
             weights: (0..node_counts.len() - 1)
@@ -29,7 +29,7 @@ impl NeuralNet {
         }
     }
 
-    pub fn feedforward(&mut self, inputs: Vec<f32>) -> Vec<f32> {
+    pub fn feedforward(&mut self, inputs: Vec<f64>) -> Vec<f64> {
         let num_layers = self.layers.len();
         self.layers[0].value = DMatrix::from_row_slice(self.layers[0].node_count, 1, &inputs);
 
@@ -49,7 +49,7 @@ impl NeuralNet {
 
 pub struct Layer {
     pub node_count: usize,
-    pub value: DMatrix<f32>,
+    pub value: DMatrix<f64>,
 }
 
 impl Layer {
@@ -61,7 +61,8 @@ impl Layer {
     }
 }
 
-fn gen_random_matrix(rows: usize, cols: usize, rng: &mut impl Rng) -> DMatrix<f32> {
+
+fn gen_random_matrix(rows: usize, cols: usize, rng: &mut impl Rng) -> DMatrix<f64> {
     let elements = rows * cols;
     let range = Uniform::new_inclusive(-1.0, 1.0);
     DMatrix::from_iterator(rows, cols, (0..elements).map(|_| range.sample(rng)))
